Make a test_helper to help clean up tests
* More tests will be added in the future (such as regression tests), so
	moving all the test overhead into a test helper is useful
